Update Hose.cs
SetupCard is run whenever a card is visualized (hovered in the card bar/displayed for picking), and twice when it's built.

Avoid usage of `*=`, `-=`, `+=, etc. in it as this means that `gun.projectileSpeed *= .66f;` is actually doing `gun.projectileSpeed = Mathf.pow(.66f, timesCardHasBeenDisplayed);` instead of `gun.projectile = 0.66f;` which causes its values to greatly differ from the intended ones.